Added implementation for GEPs courses and moved code to separate file
